removed tampermonkey errors and improved username recognition
tampermonkey showed errors when a variable declaration wasn't terminated by a semicolon.
tampermonkey also marked whitespaces at the end of a line red.
so i added semicolons after every variable declaration and removed whitespaces at the end of a line.

i also improved the recognition of usernames.
the chat webpage sometimes adds characters (zero-width whitespace) or replaces the character '-' with a nonbreaking hyphen.
these are either removed or replaced now.
hyphens and dots can be part of a username but weren't taken into account.
so i improved the regex for it.